[EI-195] [web] feat: Add convertIndicatorBoardMetadata function and updateIndicatorIdsWithSessionIds test
