+++ web/__tests__/hooks/indicator-board-metadata/use-indicator-board-metadata-view-model.hook.test.tsx
@@ -59,4 +59,27 @@ describe('useIndicatorBoardMetadataViewModel', () => {
     // then
     expect(result.current.indicatorBoardMetadata).toBeUndefined();
   });
+  describe('updateIndicatorIdsWithSessionIds', () => {
+    it('메타데이터의 세션별 지표를 변경하면, 메타데이터 값에 변경된 세션별 지표가 반영된다.', async () => {
+      // given
+      const { result } = renderHook(() => {
+        return {
+          ...useIndicatorBoardMetadataViewModel('1'),
+          ...useWorkspaceStore(),
+        };
+      });
+      await waitFor(() => expect(result.current.indicatorBoardMetadata).not.toBeUndefined());
+
+      // when
+      act(() => {
+        result.current.updateIndicatorIdsWithSessionIds({
+          session1: ['1'],
+        });
+      });
+      await waitFor(() => expect(result.current.indicatorBoardMetadata).not.toBeUndefined());
+
+      // then
+      expect(result.current.indicatorBoardMetadata?.indicatorIdsWithSessionIds['session1'][0]).toBe('1');
+    });
+  });
 });
