+++ web/app/business/hooks/indicator-board-metedata/use-indicator-board-metadata-view-model.hook.ts
@@ -2,13 +2,16 @@ import {
   IndicatorBoardMetadataResponse,
   useFetchIndicatorBoardMetadataList,
   useUpdateIndicatorBoardMetadata,
+  useUpdateIndicatorIdsWithSessionIds,
 } from '@/app/store/querys/numerical-guidance/indicator-board-metadata.query';
 import { convertIndcatorBoardMetadataList } from '../../services/view-model/indicator-board-metadata-view-model.service';
 import { useMemo } from 'react';
 
 export const useIndicatorBoardMetadataViewModel = (metadataId: string | undefined) => {
   const { data: indicatorBoardMetadataList } = useFetchIndicatorBoardMetadataList();
+
   const { trigger: updateIndicatorBoardMetadataTrigger } = useUpdateIndicatorBoardMetadata(metadataId);
+  const { trigger: updateIndicatorIdsWithSessionIdsTrigger } = useUpdateIndicatorIdsWithSessionIds(metadataId);
 
   const convertedIndicatorBoardMetadataList = useMemo(() => {
     if (!indicatorBoardMetadataList) return undefined;
@@ -35,8 +38,24 @@ export const useIndicatorBoardMetadataViewModel = (metadataId: string | undefine
     );
   };
 
+  const updateIndicatorIdsWithSessionIds = (data: { [sessionId: string]: string[] }) => {
+    updateIndicatorIdsWithSessionIdsTrigger(
+      {
+        indicatorIdsWithSessionIds: data,
+      },
+      {
+        optimisticData: (): IndicatorBoardMetadataResponse[] | undefined => {
+          convertedIndicatorBoardMetadataList?.updateIndicatorIdsWithSessionIds(metadataId, data);
+          return convertedIndicatorBoardMetadataList?.formattedIndicatorBoardMetadataList;
+        },
+        revalidate: false,
+      },
+    );
+  };
+
   return {
     indicatorBoardMetadata,
     updateIndicatorBoardMetadata,
+    updateIndicatorIdsWithSessionIds,
   };
 };
